[red-knot] Use `web-time` instead of `FileTime::now`
