+++ crates/ruff_db/src/system.rs
@@ -7,6 +7,7 @@ pub use os::testing::UserConfigDirectoryOverrideGuard;
 #[cfg(feature = "os")]
 pub use os::OsSystem;
 
+use filetime::FileTime;
 use ruff_notebook::{Notebook, NotebookError};
 use std::error::Error;
 use std::fmt::{Debug, Formatter};
@@ -345,4 +346,28 @@ impl From<glob::GlobError> for GlobError {
 pub enum GlobErrorKind {
     IOError(io::Error),
     NonUtf8Path,
+}
+
+#[cfg(not(target_arch = "wasm32"))]
+pub fn file_time_now() -> FileTime {
+    FileTime::now()
+}
+
+#[cfg(target_arch = "wasm32")]
+pub fn file_time_now() -> FileTime {
+    // Copied from FileTime::from_system_time()
+    let time = web_time::SystemTime::now();
+
+    time.duration_since(web_time::UNIX_EPOCH)
+        .map(|d| FileTime::from_unix_time(d.as_secs() as i64, d.subsec_nanos()))
+        .unwrap_or_else(|e| {
+            let until_epoch = e.duration();
+            let (sec_offset, nanos) = if until_epoch.subsec_nanos() == 0 {
+                (0, 0)
+            } else {
+                (-1, 1_000_000_000 - until_epoch.subsec_nanos())
+            };
+
+            FileTime::from_unix_time(-(until_epoch.as_secs() as i64) + sec_offset, nanos)
+        })
 }
